+++ Models/AccountViewModels/LoginRegisterModel.cs
@@ -1,8 +1,23 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.ComponentModel.DataAnnotations;
 
 namespace diary.Models.AccountViewModels
 {
+
+      public class LoginViewModel
+      {
+            [Required]
+            public string UserName { get; set; }
+
+            [Required]
+            [DataType(DataType.Password)]
+            public string Password { get; set; }
+
+            public bool RememberMe { get; set; }
+      }
+
       public class RegisterViewModel
       {
             [Required]
@@ -29,5 +44,11 @@ namespace diary.Models.AccountViewModels
 
             [Required]
             public DateTime Birthday { get; set; }
+      }
+
+      public class LoginRegisterModel
+      {
+            public LoginViewModel login { get; set; }
+            public RegisterViewModel register { get; set; }
       }
 }
