--- Models/AccountViewModels/LoginViewModel.cs
@@ -1,19 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.ComponentModel.DataAnnotations;
-
-namespace diary.Models.AccountViewModels
-{
-      public class LoginViewModel
-      {
-            [Required]
-            public string UserName { get; set; }
-
-            [Required]
-            [DataType(DataType.Password)]
-            public string Password { get; set; }
-
-            public bool RememberMe { get; set; }
-      }
-}
