--- Views/Account/Login.cshtml
@@ -1,100 +1,0 @@
-﻿@using System.Collections.Generic
-@using System.Linq
-@using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
-@model LoginViewModel
-@inject SignInManager<User> SignInManager
-
-
-@{
-    Layout = null;
-}
-
-
-<!DOCTYPE html>
-<html>
-<head>
-
-    <title>Welcome to My Diary</title>
-    <link rel="stylesheet" href="css/style.css">
-    <link href="css/popup-box.css" rel="stylesheet" type="text/css" media="all" />
-    <link href='//fonts.googleapis.com/css?family=Open+Sans:400,300italic,300,400italic,600,600italic,700,700italic,800,800italic' rel='stylesheet' type='text/css'>
-    <link href='//fonts.googleapis.com/css?family=Montserrat:400,700' rel='stylesheet' type='text/css'>
-
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    <meta name="keywords" content="Sign In And Sign Up Forms  Widget Responsive, Login Form Web Template, Flat Pricing Tables, Flat Drop-Downs, Sign-Up Web Templates, Flat Web Templates, Login Sign-up Responsive Web Template, Smartphone Compatible Web Template, Free Web Designs for Nokia, Samsung, LG, Sony Ericsson, Motorola Web Design" />
-    <script type="application/x-javascript"> addEventListener("load", function() { setTimeout(hideURLbar, 0); }, false); function hideURLbar(){ window.scrollTo(0,1); } </script>
-    <script src="js/jquery.min.js"></script>
-    <script src="js/jquery.magnific-popup.js" type="text/javascript"></script>
-    <script type="text/javascript" src="js/modernizr.custom.53451.js"></script>
-    <script>
-        $(document).ready(function() {
-            $('.popup-with-zoom-anim').magnificPopup({
-                type: 'inline',
-                fixedContentPos: false,
-                fixedBgPos: true,
-                overflowY: 'auto',
-                closeBtnInside: true,
-                preloader: false,
-                midClick: true,
-                removalDelay: 300,
-                mainClass: 'my-mfp-zoom-in'
-            });
-        });
-    </script>
-
-</head>
-<body>
-    <h1>Welcome to My Diary</h1>
-    <div class="w3layouts">
-        <div class="signin-agile">
-            <h2>Sign In</h2>
-            @using (Html.BeginForm("Login", "Account"))
-            {
-                <div asp-validation-summary="All" class="text-danger"></div>
-                <input type="text" asp-for="UserName" class="name" placeholder="Username" required="">
-                <span asp-validation-for="UserName" class="text-danger"></span>
-                <input type="password" asp-for="Password" class="password" placeholder="Password" required="">
-                <span asp-validation-for="Password" class="text-danger"></span>
-                <ul>
-                    <label asp-for="RememberMe">
-                        <input asp-for="RememberMe" type="checkbox"/>
-                        @Html.DisplayNameFor(m => m.RememberMe)
-                    </label>
-                </ul>
-                <a href="#">Forgot Password?</a><br>
-                <div class="clear"></div>
-                <input type="submit" value="Login">
-            }
-        </div>
-        <div class="signup-agileinfo">
-            <h3>Sign Up</h3>
-            <p>Not registered?<br>Sign up for free</p>
-            <div class="more">
-                <a class="book popup-with-zoom-anim button-isi zoomIn animated" data-wow-delay=".5s" href="#small-dialog">Sign Up</a>
-            </div>
-        </div>
-        <div class="clear"></div>
-    </div>
-    <div class="footer-w3l">
-        <p class="agileinfo"> © 2017 My Diary</p>
-    </div>
-    <div class="pop-up">
-        <div id="small-dialog" class="mfp-hide book-form">
-            <h3>Sign Up Form </h3>
-            <form action="#" method="post">
-                <input type="text" name="Name" placeholder="Your Name" required="" />
-                <input type="text" name="Email" class="email" placeholder="Email" required="" />
-                <input type="password" name="Password" class="password" placeholder="Password" required="" />
-                <input type="password" name="Password" class="password" placeholder="Confirm Password" required="" />
-                <input type="submit" value="Sign Up">
-            </form>
-        </div>
-    </div>
-
-    @section Scripts {
-        @await Html.PartialAsync("_ValidationScriptsPartial")
-    }
-</body>
-</html>
